requirements.txt
Django>=3.0,<4.0
django-filter>=21.1
djangorestframework>=3.12.4
Markdown>=3.3.6
psycopg2-binary>=2.8
